requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.5
Django==6.0.3
django-allauth==65.15.0
django-cloudinary-storage==0.3.0
idna==3.11
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
